# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AgencyOS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks (optionally filtered by project/assignee)
      parameters:
        - in: query
          name: projectId
          schema:
            type: string
        - in: query
          name: assigneeId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task (incl. status change, drag-drop)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Content Calendar ─────────────────────────────────────────
...
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List proposals
      parameters:
        - in: query
          name: clientId
          schema:
            type: string
      responses:
        "200":
          description: Proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Create a proposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get proposal by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
    patch:
      operationId: updateProposal
      tags: [proposals]
      summary: Update proposal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
    delete:
      operationId: deleteProposal
      tags: [proposals]
      summary: Delete proposal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Settings ─────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ─── User ─────────────────────────────────────────────
    User:
      type: object
      required: [id, email, name, systemRole, isActive, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        systemRole:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
...
    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        projectId:
          type: ["string", "null"]
        projectName:
          type: ["string", "null"]
        assigneeId:
          type: ["string", "null"]
        assigneeName:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        projectId:
          type: string
        assigneeId:
          type: string
        dueDate:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        projectId:
          type: string
        assigneeId:
          type: string
        dueDate:
          type: string

    # ─── Content Post ─────────────────────────────────────
...
    Proposal:
      type: object
      required: [id, title, clientId, status, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        clientId:
          type: string
        clientName:
          type: ["string", "null"]
        status:
          type: string
        content:
          type: ["string", "null"]
        template:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ProposalInput:
      type: object
      required: [title, clientId]
      properties:
        title:
          type: string
        clientId:
          type: string
        content:
          type: string
        template:
          type: string
    ProposalUpdate:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
        content:
          type: string
        template:
          type: string

    # ─── Settings ─────────────────────────────────────────
...